@@ -3,14 +3,13 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@apollo/client": "^3.6.0",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
     "@testing-library/user-event": "^13.5.0",
-    "apollo-boost": "^0.4.9",
-    "graphql": "^14.7.0",
-    "react": "^16.8.0",
-    "react-apollo": "^3.1.5",
-    "react-dom": "^16.8.0",
+    "graphql": "^16.4.0",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
